Revert "Revert "feat(dev/activity): added an activity collection in database.""
