+++ worktrolly-ui/src/app/body/task-details/task-details.component.ts
@@ -6,6 +6,7 @@ import { Observable } from 'rxjs';
 import { map } from 'rxjs/operators'
 import { Tasks } from 'src/app/Interface/TasksInterface';
 import { AuthService } from 'src/app/services/auth.service';
+import { ToolsService } from '../../services/tools.service';
 import {Location} from '@angular/common';
 import { NavbarHandlerService } from 'src/app/services/navbar-handler.service';
 
@@ -24,12 +25,18 @@ export class TaskDetailsComponent implements OnInit {
   editTaskEnabled: boolean = false
   userLoggedIn: boolean = false
   task: Tasks
+  todayDate: string
+  time: string
+  
   public taskDocument: AngularFirestoreDocument<Tasks>
   public taskDataObservable: Observable<Tasks>
 
-  constructor(private route: ActivatedRoute, public db: AngularFirestore, private router: Router, private functions: AngularFireFunctions, public authService: AuthService, private location: Location, private navbarHandler: NavbarHandlerService) { }
+  constructor(private route: ActivatedRoute, public db: AngularFirestore, private router: Router, private functions: AngularFireFunctions, public authService: AuthService, private location: Location, public toolsService: ToolsService, private navbarHandler: NavbarHandlerService) { }
 
   ngOnInit(): void {
+    this.todayDate = this.toolsService.date();
+    this.time = this.toolsService.time();
+
     this.Id = this.route.snapshot.params['taskId'];
 
     this.componentName = this.Id;
@@ -70,7 +77,7 @@ export class TaskDetailsComponent implements OnInit {
     const callable = this.functions.httpsCallable('deleteTask');
 
     try {
-      const result = await callable({ Id: this.task.Id, SprintNumber: this.task.SprintNumber, Category: this.task.Category, Status: this.task.Status }).toPromise();
+      const result = await callable({ Id: this.task.Id, SprintNumber: this.task.SprintNumber, Category: this.task.Category, Status: this.task.Status, Date: this.todayDate, Time: this.time }).toPromise();
       console.log(this.task.Id + " deleted");
       console.log(result);
       this.router.navigate(['/']);
